Fixed error with cast timezone field to double.

+++ VkToolkit/Friends.cs
@@ -26,7 +26,7 @@ namespace VkToolkit
         /// <param name="offset">Recors offset</param>
         /// <param name="order">Friends order(only for standalone desktop applications) </param>
         /// <returns>List of users</returns>
-        public IEnumerable<Profile> Get(int uid, ProfileFields fields = null, int? count = null, int? offset = null, Order order = null)
+        public IEnumerable<Profile> Get(long uid, ProfileFields fields = null, int? count = null, int? offset = null, Order order = null)
         {
             _vk.IfAccessTokenNotDefinedThrowException();
